requirements.txt
torch
torchvision
pillow
numpy
pandas
tqdm
datasets
jupyter
ipykernel
sentence_transformers
faiss-cpu